Updated unit test for ClientTypingActivity
