+++ ProjectSource/app-library/src/test/java/com/kayako/sdk/android/k5/messenger/data/conversation/viewmodel/ClientTypingActivityTest.java
@@ -1,28 +1,21 @@
 package com.kayako.sdk.android.k5.messenger.data.conversation.viewmodel;
 
+import com.kayako.sdk.android.k5.EqualsAndHashTest;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ErrorCollector;
 import org.junit.rules.ExpectedException;
 import java.util.Map;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.containsString;
 
-public class ClientTypingActivityTest {
+public class ClientTypingActivityTest extends EqualsAndHashTest<ClientTypingActivity>{
 
     private static final boolean IS_TYPING = Boolean.TRUE;
     private UserViewModel user;
-    private ClientTypingActivity one;
-    private ClientTypingActivity same;
-    private ClientTypingActivity secondSame;
-    private ClientTypingActivity other;
 
     @Rule
     public final ExpectedException thrown = ExpectedException.none();
-
-    @Rule
-    public final ErrorCollector errorCollector = new ErrorCollector();
 
     @Before
     public void setUp() {
@@ -38,65 +31,54 @@ public class ClientTypingActivityTest {
 
     @Test
     public void whenValidParamsConstructorThenObjectCreated() {
+        //Assert
         errorCollector.checkThat(one.isTyping(), is(IS_TYPING));
         errorCollector.checkThat(one.getUser(), is(user));
     }
 
     @Test
     public void whenIsTypingFalseThenObjectCreated() {
+        //Act
         ClientTypingActivity clientTypingActivityLocal = new ClientTypingActivity(false);
+
+        //Assert
         errorCollector.checkThat(clientTypingActivityLocal.isTyping(), is(false));
     }
 
     @Test
     public void whenIsTypingTrueThenIllegalArgumentException() {
+        //Arrange
         final String exceptionMessage =
             "If isTyping is true, then user  must also be provided. Use the appropriate constructor";
+
+        //Contract
         thrown.expect(IllegalArgumentException.class);
         thrown.expectMessage(containsString(exceptionMessage));
+
+        //Act
         new ClientTypingActivity(IS_TYPING);
     }
 
     @Test
     public void whenUserIsNullThenIllegalArgumentException() {
+        //Arrange
         final String exceptionMessage =
                 "If isTyping is true, then user  must also be provided!";
+
+        //Contract
         thrown.expect(IllegalArgumentException.class);
         thrown.expectMessage(containsString(exceptionMessage));
+
+        //Act
         new ClientTypingActivity(IS_TYPING, null);
     }
 
     @Test
-    public void reflexivity() {
-        errorCollector.checkThat(one, is(one));
-    }
-
-    @Test
-    public void nullInequality() {
-        errorCollector.checkThat(one.equals(null), is(false));
-    }
-
-    @Test
-    public void symmetry() {
-        errorCollector.checkThat(one, is(same));
-        errorCollector.checkThat(same, is(one));
-        errorCollector.checkThat(one.hashCode(), is(same.hashCode()));
-        errorCollector.checkThat(one.equals(other), is(false));
-        errorCollector.checkThat(other.equals(one), is(false));
-    }
-
-    @Test
-    public void transitivity() {
-        errorCollector.checkThat(one, is(same));
-        errorCollector.checkThat(same, is(secondSame));
-        errorCollector.checkThat(one, is(secondSame));
-        errorCollector.checkThat(one.hashCode(), is(same.hashCode()));
-        errorCollector.checkThat(one.hashCode(), is(secondSame.hashCode()));
-    }
-
-    @Test
     public void getContents() {
+        //Act
         Map<String, String > contentsMap = one.getContents();
+
+        //Assert
         errorCollector.checkThat(contentsMap.get("isTyping"), is(String.valueOf(IS_TYPING)));
     }
 }
